src/helena-dialect/namespaces.spec.ts
import { expect } from "chai";
import {
  BREAK,
  CONTINUE,
  ERROR,
  OK,
  ResultCode,
  RETURN,
} from "../core/results";
import { Parser } from "../core/parser";
import { Tokenizer } from "../core/tokenizer";
import { StringValue } from "../core/values";
import { CommandValue, Scope, Variable } from "./core";
import { initCommands } from "./helena-dialect";

describe("Helena namespaces", () => {
  let rootScope: Scope;

  let tokenizer: Tokenizer;
  let parser: Parser;

  const parse = (script: string) =>
    parser.parse(tokenizer.tokenize(script)).script;
  const execute = (script: string) => rootScope.executeScript(parse(script));
  const evaluate = (script: string) => execute(script).value;

  beforeEach(() => {
    rootScope = new Scope();
    initCommands(rootScope);

    tokenizer = new Tokenizer();
    parser = new Parser();
  });

  describe("namespace", () => {
    it("should define a new command", () => {
      evaluate("namespace cmd {}");
      expect(rootScope.context.commands.has("cmd")).to.be.true;
    });
    it("should replace existing commands", () => {
      evaluate("namespace cmd {}");
      expect(execute("namespace cmd {}").code).to.eql(ResultCode.OK);
    });
    it("should return a command value", () => {
      expect(evaluate("namespace {}")).to.be.instanceof(CommandValue);
      expect(evaluate("namespace cmd {}")).to.be.instanceof(CommandValue);
    });
    specify("command should return self", () => {
      const value = evaluate("namespace cmd {}");
      expect(evaluate("cmd")).to.eql(value);
    });
    specify("command value should return self", () => {
      const value = evaluate("set cmd [namespace {}]");
      expect(evaluate("$cmd")).to.eql(value);
    });
    describe("body", () => {
      it("should be executed", () => {
        evaluate("closure cmd {} {let var val}");
        expect(rootScope.context.constants.has("var")).to.be.false;
        evaluate("namespace {cmd}");
        expect(rootScope.context.constants.get("var")).to.eql(
          new StringValue("val")
        );
      });
      it("should access global commands", () => {
        expect(execute("namespace {idem val}").code).to.eql(ResultCode.OK);
      });
      it("should not access global variables", () => {
        evaluate("set var val");
        expect(execute("namespace {get var}").code).to.eql(ResultCode.ERROR);
      });
      it("should not set global variables", () => {
        evaluate("set var val");
        evaluate("namespace {set var val2; let cst val3}");
        expect(rootScope.context.variables.get("var")).to.eql(
          new Variable(new StringValue("val"))
        );
        expect(rootScope.context.constants.has("cst")).to.be.false;
      });
      it("should set namespace variables", () => {
        evaluate("set var val");
        evaluate("namespace cmd {set var val2; let cst val3}");
        expect(rootScope.context.variables.get("var")).to.eql(
          new Variable(new StringValue("val"))
        );
        expect(rootScope.context.constants.has("cst")).to.be.false;
        expect(evaluate("[cmd] eval {get var}")).to.eql(
          new StringValue("val2")
        );
        expect(evaluate("[cmd] eval {get cst}")).to.eql(
          new StringValue("val3")
        );
      });
      describe("exceptions", () => {
        specify("non-script body", () => {
          expect(execute("namespace a")).to.eql(ERROR("body must be a script"));
          expect(execute("namespace a b")).to.eql(
            ERROR("body must be a script")
          );
        });
      });
    });
    describe("control flow", () => {
      describe("return", () => {
        it("should interrupt the body with OK code", () => {
          evaluate("closure cmd1 {} {set var val1}");
          evaluate("closure cmd2 {} {set var val2}");
          expect(execute("namespace {cmd1; return; cmd2}").code).to.eql(
            ResultCode.OK
          );
          expect(evaluate("get var")).to.eql(new StringValue("val1"));
        });
        it("should still define the namespace command", () => {
          evaluate("namespace cmd {return}");
          expect(rootScope.context.commands.has("cmd")).to.be.true;
        });
        it("should return passed value instead of namespace command value", () => {
          expect(execute("namespace {return val}")).to.eql(
            OK(new StringValue("val"))
          );
        });
      });
      describe("tailcall", () => {
        it("should interrupt the body with OK code", () => {
          evaluate("closure cmd1 {} {set var val1}");
          evaluate("closure cmd2 {} {set var val2}");
          expect(execute("namespace {cmd1; tailcall {}; cmd2}").code).to.eql(
            ResultCode.OK
          );
          expect(evaluate("get var")).to.eql(new StringValue("val1"));
        });
        it("should still define the namespace command", () => {
          evaluate("namespace cmd {tailcall {}}");
          expect(rootScope.context.commands.has("cmd")).to.be.true;
        });
        it("should return passed value instead of namespace command value", () => {
          expect(execute("namespace {tailcall {idem val}}")).to.eql(
            OK(new StringValue("val"))
          );
        });
      });
      describe("yield", () => {
        it("should interrupt the body with YIELD code", () => {
          evaluate("closure cmd1 {} {set var val1}");
          evaluate("closure cmd2 {} {set var val2}");
          expect(execute("namespace cmd {cmd1; yield; cmd2}").code).to.eql(
            ResultCode.YIELD
          );
          expect(evaluate("get var")).to.eql(new StringValue("val1"));
        });
        it("should provide a resumable state", () => {
          evaluate("closure cmd1 {} {set var val1}");
          evaluate("closure cmd2 {val} {set var $val}");
          const process = rootScope.prepareScript(
            parse("namespace cmd {cmd1; cmd2 _[yield val2]_}")
          );

          let result = process.run();
          expect(result.code).to.eql(ResultCode.YIELD);
          expect(result.value).to.eql(new StringValue("val2"));
          expect(result.data).to.exist;

          process.yieldBack(new StringValue("val3"));
          result = process.run();
          expect(result.code).to.eql(ResultCode.OK);
          expect(result.value).to.be.instanceof(CommandValue);
          expect(evaluate("get var")).to.eql(new StringValue("_val3_"));
        });
        it("should delay the definition of namespace command until resumed", () => {
          const process = rootScope.prepareScript(
            parse("namespace cmd {yield}")
          );

          let result = process.run();
          expect(result.code).to.eql(ResultCode.YIELD);
          expect(rootScope.context.commands.has("cmd")).to.be.false;

          result = process.run();
          expect(result.code).to.eql(ResultCode.OK);
          expect(rootScope.context.commands.has("cmd")).to.be.true;
        });
      });
      describe("error", () => {
        it("should interrupt the body with ERROR code", () => {
          evaluate("closure cmd1 {} {set var val1}");
          evaluate("closure cmd2 {} {set var val2}");
          expect(execute("namespace {cmd1; error msg; cmd2}")).to.eql(
            ERROR("msg")
          );
          expect(evaluate("get var")).to.eql(new StringValue("val1"));
        });
        it("should not define the namespace command", () => {
          evaluate("namespace cmd {error msg}");
          expect(rootScope.context.commands.has("cmd")).to.be.false;
        });
      });
      describe("break", () => {
        it("should interrupt the body with ERROR code", () => {
          evaluate("closure cmd1 {} {set var val1}");
          evaluate("closure cmd2 {} {set var val2}");
          expect(execute("namespace {cmd1; break; cmd2}")).to.eql(
            ERROR("unexpected break")
          );
          expect(evaluate("get var")).to.eql(new StringValue("val1"));
        });
        it("should not define the namespace command", () => {
          evaluate("namespace cmd {break}");
          expect(rootScope.context.commands.has("cmd")).to.be.false;
        });
      });
      describe("continue", () => {
        it("should interrupt the body with ERROR code", () => {
          evaluate("closure cmd1 {} {set var val1}");
          evaluate("closure cmd2 {} {set var val2}");
          expect(execute("namespace {cmd1; continue; cmd2}")).to.eql(
            ERROR("unexpected continue")
          );
          expect(evaluate("get var")).to.eql(new StringValue("val1"));
        });
        it("should not define the namespace command", () => {
          evaluate("namespace cmd {continue}");
          expect(rootScope.context.commands.has("cmd")).to.be.false;
        });
      });
    });
    describe("methods", () => {
      describe("eval", () => {
        it("should evaluate body in namespace scope", () => {
          evaluate("namespace cmd {let cst val}");
          expect(evaluate("[cmd] eval {get cst}")).to.eql(
            new StringValue("val")
          );
        });
        it("should accept tuple bodies", () => {
          evaluate("namespace cmd {let cst val}");
          expect(evaluate("[cmd] eval (get cst)")).to.eql(
            new StringValue("val")
          );
        });
        it("should evaluate macros in namespace scope", () => {
          evaluate("namespace cmd {macro mac {} {let cst val}}");
          evaluate("[cmd] eval {mac}");
          expect(rootScope.context.constants.has("cst")).to.be.false;
          expect(evaluate("[cmd] eval {get cst}")).to.eql(
            new StringValue("val")
          );
        });
        it("should evaluate closures in their scope", () => {
          evaluate("closure cls {} {let cst val}");
          evaluate("namespace cmd {}");
          evaluate("[cmd] eval {cls}");
          expect(rootScope.context.constants.get("cst")).to.eql(
            new StringValue("val")
          );
          expect(execute("[cmd] eval {get cst}").code).to.eql(ResultCode.ERROR);
        });
        describe("control flow", () => {
          describe("return", () => {
            it("should interrupt the body with RETURN code", () => {
              evaluate("closure cmd1 {} {set var val1}");
              evaluate("closure cmd2 {} {set var val2}");
              evaluate("namespace cmd {}");
              expect(execute("[cmd] eval {cmd1; return val3; cmd2}")).to.eql(
                RETURN(new StringValue("val3"))
              );
              expect(evaluate("get var")).to.eql(new StringValue("val1"));
            });
          });
          describe("tailcall", () => {
            it("should interrupt the body with RETURN code", () => {
              evaluate("closure cmd1 {} {set var val1}");
              evaluate("closure cmd2 {} {set var val2}");
              evaluate("namespace cmd {}");
              expect(
                execute("[cmd] eval {cmd1; tailcall {idem val3}; cmd2}")
              ).to.eql(RETURN(new StringValue("val3")));
              expect(evaluate("get var")).to.eql(new StringValue("val1"));
            });
          });
          describe("yield", () => {
            it("should interrupt the body with YIELD code", () => {
              evaluate("closure cmd1 {} {set var val1}");
              evaluate("closure cmd2 {} {set var val2}");
              evaluate("namespace cmd {}");
              expect(execute("[cmd] eval {cmd1; yield; cmd2}").code).to.eql(
                ResultCode.YIELD
              );
              expect(evaluate("get var")).to.eql(new StringValue("val1"));
            });
            it("should provide a resumable state", () => {
              evaluate("closure cmd1 {} {set var val1}");
              evaluate("closure cmd2 {val} {set var $val}");
              evaluate("namespace cmd {}");
              const process = rootScope.prepareScript(
                parse("[cmd] eval {cmd1; cmd2 _[yield val2]_}")
              );

              let result = process.run();
              expect(result.code).to.eql(ResultCode.YIELD);
              expect(result.value).to.eql(new StringValue("val2"));

              process.yieldBack(new StringValue("val3"));
              result = process.run();
              expect(result).to.eql(OK(new StringValue("_val3_")));
              expect(evaluate("get var")).to.eql(new StringValue("_val3_"));
            });
          });
          describe("error", () => {
            it("should interrupt the body with ERROR code", () => {
              evaluate("closure cmd1 {} {set var val1}");
              evaluate("closure cmd2 {} {set var val2}");
              evaluate("namespace cmd {}");
              expect(execute("[cmd] eval {cmd1; error msg; cmd2}")).to.eql(
                ERROR("msg")
              );
              expect(evaluate("get var")).to.eql(new StringValue("val1"));
            });
          });
          describe("break", () => {
            it("should interrupt the body with BREAK code", () => {
              evaluate("closure cmd1 {} {set var val1}");
              evaluate("closure cmd2 {} {set var val2}");
              evaluate("namespace cmd {}");
              expect(execute("[cmd] eval {cmd1; break; cmd2}")).to.eql(BREAK());
              expect(evaluate("get var")).to.eql(new StringValue("val1"));
            });
          });
          describe("continue", () => {
            it("should interrupt the body with CONTINUE code", () => {
              evaluate("closure cmd1 {} {set var val1}");
              evaluate("closure cmd2 {} {set var val2}");
              evaluate("namespace cmd {}");
              expect(execute("[cmd] eval {cmd1; continue; cmd2}")).to.eql(
                CONTINUE()
              );
              expect(evaluate("get var")).to.eql(new StringValue("val1"));
            });
          });
        });
        describe("exceptions", () => {
          specify("wrong arity", () => {
            expect(execute("[namespace {}] eval")).to.eql(
              ERROR('wrong # args: should be "namespace eval body"')
            );
            expect(execute("[namespace {}] eval a b")).to.eql(
              ERROR('wrong # args: should be "namespace eval body"')
            );
          });
          specify("invalid body", () => {
            expect(execute("[namespace {}] eval 1")).to.eql(
              ERROR("body must be a script or tuple")
            );
          });
        });
      });
      describe("call", () => {
        it("should call namespace commands", () => {
          evaluate("namespace cmd {macro mac {} {idem val}}");
          expect(evaluate("[cmd] call mac")).to.eql(new StringValue("val"));
        });
        it("should evaluate macros in namespace", () => {
          evaluate("namespace cmd {macro mac {} {let cst val}}");
          evaluate("[cmd] call mac");
          expect(rootScope.context.constants.has("cst")).to.be.false;
          expect(evaluate("[cmd] eval {get cst}")).to.eql(
            new StringValue("val")
          );
        });
        it("should evaluate namespace closures in namespace", () => {
          evaluate("namespace cmd {closure cls {} {let cst val}}");
          evaluate("[cmd] call cls");
          expect(rootScope.context.constants.has("cst")).to.be.false;
          expect(evaluate("[cmd] eval {get cst}")).to.eql(
            new StringValue("val")
          );
        });
        describe("control flow", () => {
          describe("return", () => {
            it("should interrupt the body with RETURN code", () => {
              evaluate("closure cmd1 {} {set var val1}");
              evaluate("closure cmd2 {} {set var val2}");
              evaluate(
                "namespace cmd {macro mac {} {cmd1; return val3; cmd2}}"
              );
              expect(execute("[cmd] call mac")).to.eql(
                RETURN(new StringValue("val3"))
              );
              expect(evaluate("get var")).to.eql(new StringValue("val1"));
            });
          });
          describe("tailcall", () => {
            it("should interrupt the body with RETURN code", () => {
              evaluate("closure cmd1 {} {set var val1}");
              evaluate("closure cmd2 {} {set var val2}");
              evaluate(
                "namespace cmd {macro mac {} {cmd1; tailcall {idem val3}; cmd2}}"
              );
              expect(execute("[cmd] call mac")).to.eql(
                RETURN(new StringValue("val3"))
              );
              expect(evaluate("get var")).to.eql(new StringValue("val1"));
            });
          });
          describe("yield", () => {
            it("should interrupt the call with YIELD code", () => {
              evaluate("closure cmd1 {} {set var val1}");
              evaluate("closure cmd2 {} {set var val2}");
              evaluate("namespace cmd {macro mac {} {cmd1; yield; cmd2}}");
              expect(execute("[cmd] call mac").code).to.eql(ResultCode.YIELD);
              expect(evaluate("get var")).to.eql(new StringValue("val1"));
            });
            it("should provide a resumable state", () => {
              evaluate("closure cmd1 {} {set var val1}");
              evaluate("closure cmd2 {val} {set var $val}");
              evaluate("namespace cmd {proc p {} {cmd1; cmd2 _[yield val2]_}}");
              const process = rootScope.prepareScript(parse("[cmd] call p"));

              let result = process.run();
              expect(result.code).to.eql(ResultCode.YIELD);
              expect(result.value).to.eql(new StringValue("val2"));

              process.yieldBack(new StringValue("val3"));
              result = process.run();
              expect(result).to.eql(OK(new StringValue("_val3_")));
              expect(evaluate("get var")).to.eql(new StringValue("_val3_"));
            });
          });
          describe("error", () => {
            it("should interrupt the body with ERROR code", () => {
              evaluate("closure cmd1 {} {set var val1}");
              evaluate("closure cmd2 {} {set var val2}");
              evaluate("namespace cmd {macro mac {} {cmd1; error msg; cmd2}}");
              expect(execute("[cmd] call mac")).to.eql(ERROR("msg"));
              expect(evaluate("get var")).to.eql(new StringValue("val1"));
            });
          });
          describe("break", () => {
            it("should interrupt the body with BREAK code", () => {
              evaluate("closure cmd1 {} {set var val1}");
              evaluate("closure cmd2 {} {set var val2}");
              evaluate("namespace cmd {macro mac {} {cmd1; break; cmd2}}");
              expect(execute("[cmd] call mac")).to.eql(BREAK());
              expect(evaluate("get var")).to.eql(new StringValue("val1"));
            });
          });
          describe("continue", () => {
            it("should interrupt the body with CONTINUE code", () => {
              evaluate("closure cmd1 {} {set var val1}");
              evaluate("closure cmd2 {} {set var val2}");
              evaluate("namespace cmd {macro mac {} {cmd1; continue; cmd2}}");
              expect(execute("[cmd] call mac")).to.eql(CONTINUE());
              expect(evaluate("get var")).to.eql(new StringValue("val1"));
            });
          });
        });
        describe("exceptions", () => {
          specify("wrong arity", () => {
            expect(execute("[namespace {}] call")).to.eql(
              ERROR(
                'wrong # args: should be "namespace call cmdname ?arg ...?"'
              )
            );
          });
          specify("non-existing command", () => {
            expect(execute("[namespace {}] call unknownCommand")).to.eql(
              ERROR('invalid command name "unknownCommand"')
            );
          });
          specify("out-of-scope command", () => {
            expect(execute("macro cmd {} {}; [namespace {}] call cmd")).to.eql(
              ERROR('invalid command name "cmd"')
            );
          });
        });
      });
      describe("exceptions", () => {
        specify("non-existing method", () => {
          expect(execute("[namespace {}] unknownMethod")).to.eql(
            ERROR('invalid method name "unknownMethod"')
          );
        });
      });
    });
    describe("subcommands", () => {
      it("should map first argument to namespace command name", () => {
        evaluate("namespace cmd {macro opt {} {idem val}}");
        expect(evaluate("cmd opt")).to.eql(new StringValue("val"));
      });
      it("should pass remaining arguments to namespace command", () => {
        evaluate("namespace cmd {macro opt {arg} {idem $arg}}");
        expect(evaluate("cmd opt val")).to.eql(new StringValue("val"));
      });
      it("should evaluate command in namespace scope", () => {
        evaluate("namespace cmd {macro mac {} {let cst val}}");
        evaluate("cmd mac");
        expect(rootScope.context.constants.has("cst")).to.be.false;
        expect(evaluate("[cmd] eval {get cst}")).to.eql(new StringValue("val"));
      });
      it("should work recursively", () => {
        evaluate("namespace ns1 {namespace ns2 {macro opt {} {idem val}}}");
        expect(evaluate("ns1 ns2 opt")).to.eql(new StringValue("val"));
      });
      describe("control flow", () => {
        describe("return", () => {
          it("should interrupt the body with RETURN code", () => {
            evaluate("closure cmd1 {} {set var val1}");
            evaluate("closure cmd2 {} {set var val2}");
            evaluate("namespace cmd {macro mac {} {cmd1; return val3; cmd2}}");
            expect(execute("cmd mac")).to.eql(RETURN(new StringValue("val3")));
            expect(evaluate("get var")).to.eql(new StringValue("val1"));
          });
        });
        describe("tailcall", () => {
          it("should interrupt the body with RETURN code", () => {
            evaluate("closure cmd1 {} {set var val1}");
            evaluate("closure cmd2 {} {set var val2}");
            evaluate(
              "namespace cmd {macro mac {} {cmd1; tailcall {idem val3}; cmd2}}"
            );
            expect(execute("cmd mac")).to.eql(RETURN(new StringValue("val3")));
            expect(evaluate("get var")).to.eql(new StringValue("val1"));
          });
        });
        describe("yield", () => {
          it("should interrupt the call with YIELD code", () => {
            evaluate("closure cmd1 {} {set var val1}");
            evaluate("closure cmd2 {} {set var val2}");
            evaluate("namespace cmd {macro mac {} {cmd1; yield; cmd2}}");
            expect(execute("cmd mac").code).to.eql(ResultCode.YIELD);
            expect(evaluate("get var")).to.eql(new StringValue("val1"));
          });
          it("should provide a resumable state", () => {
            evaluate("closure cmd1 {} {set var val1}");
            evaluate("closure cmd2 {val} {set var $val}");
            evaluate("namespace cmd {proc p {} {cmd1; cmd2 _[yield val2]_}}");
            const process = rootScope.prepareScript(parse("cmd p"));

            let result = process.run();
            expect(result.code).to.eql(ResultCode.YIELD);
            expect(result.value).to.eql(new StringValue("val2"));

            process.yieldBack(new StringValue("val3"));
            result = process.run();
            expect(result).to.eql(OK(new StringValue("_val3_")));
            expect(evaluate("get var")).to.eql(new StringValue("_val3_"));
          });
        });
        describe("error", () => {
          it("should interrupt the body with ERROR code", () => {
            evaluate("closure cmd1 {} {set var val1}");
            evaluate("closure cmd2 {} {set var val2}");
            evaluate("namespace cmd {macro mac {} {cmd1; error msg; cmd2}}");
            expect(execute("cmd mac")).to.eql(ERROR("msg"));
            expect(evaluate("get var")).to.eql(new StringValue("val1"));
          });
        });
        describe("break", () => {
          it("should interrupt the body with BREAK code", () => {
            evaluate("closure cmd1 {} {set var val1}");
            evaluate("closure cmd2 {} {set var val2}");
            evaluate("namespace cmd {macro mac {} {cmd1; break; cmd2}}");
            expect(execute("cmd mac")).to.eql(BREAK());
            expect(evaluate("get var")).to.eql(new StringValue("val1"));
          });
        });
        describe("continue", () => {
          it("should interrupt the body with CONTINUE code", () => {
            evaluate("closure cmd1 {} {set var val1}");
            evaluate("closure cmd2 {} {set var val2}");
            evaluate("namespace cmd {macro mac {} {cmd1; continue; cmd2}}");
            expect(execute("cmd mac")).to.eql(CONTINUE());
            expect(evaluate("get var")).to.eql(new StringValue("val1"));
          });
        });
      });
      describe("exceptions", () => {
        specify("non-existing command", () => {
          evaluate("namespace cmd {}");
          expect(execute("cmd unknownCommand")).to.eql(
            ERROR('invalid command name "unknownCommand"')
          );
        });
        specify("out-of-scope command", () => {
          evaluate("macro mac {} {}; namespace cmd {}");
          expect(execute("cmd mac")).to.eql(
            ERROR('invalid command name "mac"')
          );
        });
      });
    });
    describe("variables", () => {
      it("should map to value keys", () => {
        evaluate("set ns [namespace cmd {let cst val1; set var val2}]");
        expect(evaluate("idem $[cmd](cst)")).to.eql(new StringValue("val1"));
        expect(evaluate("idem $[cmd](var)")).to.eql(new StringValue("val2"));
        expect(evaluate("idem $ns(cst)")).to.eql(new StringValue("val1"));
        expect(evaluate("idem $ns(var)")).to.eql(new StringValue("val2"));
        evaluate("$ns eval {set var2 val3}");
        expect(evaluate("idem $ns(var2)")).to.eql(new StringValue("val3"));
      });
      it("should work recursively", () => {
        evaluate(
          "set ns1 [namespace {set ns2 [namespace {let cst val1; set var val2}]}]"
        );
        expect(evaluate("idem $ns1(ns2)(cst)")).to.eql(new StringValue("val1"));
        expect(evaluate("idem $ns1(ns2)(var)")).to.eql(new StringValue("val2"));
        expect(evaluate("idem $ns1(ns2 cst)")).to.eql(new StringValue("val1"));
        expect(evaluate("idem $ns1(ns2 var)")).to.eql(new StringValue("val2"));
      });
      describe("exceptions", () => {
        specify("non-existing variables", () => {
          evaluate("namespace cmd {}");
          expect(execute("$[cmd](unknownVariable)")).to.eql(
            ERROR('cannot get "unknownVariable": no such variable')
          );
        });
        specify("out-of-scope variable", () => {
          evaluate("let cst var; namespace cmd {}");
          expect(execute("$[cmd](cst)")).to.eql(
            ERROR('cannot get "cst": no such variable')
          );
        });
      });
    });
    describe("exceptions", () => {
      specify("wrong arity", () => {
        expect(execute("namespace")).to.eql(
          ERROR('wrong # args: should be "namespace ?name? body"')
        );
        expect(execute("namespace a b c")).to.eql(
          ERROR('wrong # args: should be "namespace ?name? body"')
        );
      });
    });
  });
});
